+++ doc/source/notebooks/convert.py
@@ -5,7 +5,7 @@
 #       extension: .py
 #       format_name: percent
 #       format_version: '1.3'
-#       jupytext_version: 1.16.4
+#       jupytext_version: 1.16.7
 #   kernelspec:
 #     display_name: Python 3 (ipykernel)
 #     language: python
@@ -13,7 +13,7 @@
 # ---
 
 # %% [markdown]
-# # Convert MRIO satellite raws
+# # Convert MRIO satellite rows
 
 # %% [markdown]
 # Here we discuss the possibilities for converting entries in MRIO satellite accounts (Extensions).