+++ CHANGELOG.rst
@@ -4,6 +4,23 @@ Changelog
 
 
 ******************
+v0.6.1 - dev
+******************
+
+
+New Features
+============
+
+* `convert` functions (ioutil.convert, Extension.convert, pymrio.extension_convert) have a new argument
+  `reindex` which specifies a order of the converted dataframe/extension. It is a wrapper around pandas reindex,
+  but also allows to passing of a bridge column name to sort after the order given in the bridge table.
+
+Fixes
+======
+
+* spelling mistakes in docs
+
+******************
 v0.6.0 - 20250616
 ******************